requirements.txt
python-telegram-bot
thefuzz
python-levenshtein
duckduckgo-search
